+++ cypress/e2e/spec-wc.cy.js
@@ -94,8 +94,7 @@ describe("default behaviour", () => {
 describe("when load_remix_disabled is true, e.g. in editor-standalone", () => {
   const authKey = `oidc.user:https://auth-v1.raspberrypi.org:editor-api`;
 
-  const oneHourFromNow = (new Date().getTime() + 60 * 60 * 1000) / 1000;
-  const user = { access_token: "dummy-access-token", expires_at: oneHourFromNow };
+  const user = { access_token: "dummy-access-token" };
   const originalIdentifier = "blank-python-starter";
 
   const urlFor = (identifier) => {