+++ src/containers/WebComponentLoader.jsx
@@ -48,29 +48,13 @@ const WebComponentLoader = (props) => {
     useEditorStyles = false, // If true use the standard editor styling for the web component
   } = props;
 
-  const getLocalStorageUser = (authKey) => {
-    if (authKey) {
-      const user = JSON.parse(localStorage.getItem(authKey));
-      if (user) {
-        const expiresAt = new Date(user.expires_at * 1000);
-        if (new Date() < expiresAt) {
-          return user;
-        } else {
-          return null;
-        }
-      } else {
-        return null;
-      }
-    } else {
-      return null;
-    }
-  };
-
   const dispatch = useDispatch();
   const { t } = useTranslation();
   const [projectIdentifier, setProjectIdentifier] = useState(identifier);
   localStorage.setItem("authKey", authKey);
-  const localStorageUser = getLocalStorageUser(authKey);
+  const localStorageUser = authKey
+    ? JSON.parse(localStorage.getItem(authKey))
+    : null;
   const user = useSelector((state) => state.auth.user || localStorageUser);
   const [loadCache, setLoadCache] = useState(!!!user);
   const [loadRemix, setLoadRemix] = useState(!!user);